+++ CHANGELOG.md
@@ -185,6 +185,7 @@
    * ADDED: Linear reference support to locate api [#2645](https://github.com/valhalla/valhalla/pull/2645)
    * ADDED: Implemented OSRM-like turn duration calculation for car. Uses it now in auto costing. [#2651](https://github.com/valhalla/valhalla/pull/2651)
    * ADDED: Enhanced turn lane information in guidance [#2653](https://github.com/valhalla/valhalla/pull/2653)
+   * ADDED: `top_speed` option for auto & truck []
 
 ## Release Date: 2019-11-21 Valhalla 3.0.9
 * **Bug Fix**